[focus] document current focus
Probably needs some changelog, there are modules for that, and more organisation and documentation.
Some form of calendar may be good. But first I want the [rules panel] working and [crumbs] restored and working to give a sense of what the whole project is about. And generate a .gif that shows it.